Anagram grouping task.
The previous method was redundant, I made the mistake of using a solution from an easy anagram than after looking at the video solution I've learned a valuable lesson. What in one problem works as a solution at a different problem you may find an optimal better solution and not throwing crap in hopes of working.